CAS-165 Update Create New Community css (#182)
* add migration file to update text

* update text

* update text

* remove marging and align items

* add popover component
