+++ frontend/packages/client/src/components/Community/CommunityEditorDetails.js
@@ -41,7 +41,7 @@ export const CommunityUsersForm = ({
           <div className="is-flex flex-1">
             <WrapperResponsive
               tag="h5"
-              classNames="title is-6 mb-2 is-flex"
+              classNames="title is-6 is-flex is-align-items-center"
               extraClassesMobile="mt-4"
             >
               {title}
