+++ frontend/packages/client/src/components/CommunityCreate/StepTwo.js
@@ -1,19 +1,14 @@
 import React, { useEffect } from 'react';
 import { CommunityUsersForm } from '../Community/CommunityEditorDetails';
 import useFlowAddrValidator from '../Community/hooks/useFlowAddrValidator';
-import { useMediaQuery } from 'hooks';
-import classnames from 'classnames';
+import Popover from 'components/Popover';
 
 const isInitialList = (listAddr) => {
   return listAddr?.length === 1 && listAddr[0].addr === '';
 };
 
-const buttonStyle = {
-  border: 'none',
-  fontSize: '12px',
-  marginLeft: '4px',
-};
-
+const popoverParagraph =
+  'In addition, community creator address will be set as admin and member by default.';
 export default function StepTwo({
   stepData,
   setStepValid,
@@ -21,8 +16,6 @@ export default function StepTwo({
   moveToNextStep,
   isStepValid,
 }) {
-  const notMobile = useMediaQuery();
-
   const { listAddrAdmins = [{ addr: '' }], listAddrAuthors = [{ addr: '' }] } =
     stepData || {};
 
@@ -87,49 +80,20 @@ export default function StepTwo({
     onDataChange({ listAddrAuthors: [...listAddrAuthors, { addr: '' }] });
   };
 
-  const popoverClassName = classnames(
-    'columns',
-    { 'm-4': notMobile },
-    { 'm-2': !notMobile }
-  );
   return (
     <>
       <CommunityUsersForm
         title={
           <>
             Admins
-            <div className="popover is-popover-bottom">
-              <button
-                className="delete has-text-grey rounded-full cursor-pointer popover-trigger"
-                style={buttonStyle}
-              >
-                ?
-              </button>
-              <div className="popover-content">
-                <div className={popoverClassName}>
-                  <div className="column is-12 p-0">
-                    <p
-                      className="small-text has-text-weight-normal has-text-grey small-text mb-1"
-                      style={{
-                        lineHeight: '20px',
-                      }}
-                    >
-                      Admin addresses will be added automatically as authors and
-                      members for the community.
-                    </p>
-                    <p
-                      className="small-text has-text-weight-normal has-text-grey small-text"
-                      style={{
-                        lineHeight: '20px',
-                      }}
-                    >
-                      In addition, community creator address will be set as
-                      admin and member by default.
-                    </p>
-                  </div>
-                </div>
-              </div>
-            </div>
+            <Popover
+              paragraphs={[
+                'Admin addresses will be added automatically as authors and members for the community.',
+                popoverParagraph,
+              ]}
+            >
+              ?
+            </Popover>
           </>
         }
         description="The admins will be able to edit the space settings and moderate proposals. You must add one address per line."
@@ -144,7 +108,19 @@ export default function StepTwo({
         autoFocusOnLoad={true}
       />
       <CommunityUsersForm
-        title="Authors"
+        title={
+          <>
+            Authors
+            <Popover
+              paragraphs={[
+                'Author addresses will be added automatically as members for the community.',
+                popoverParagraph,
+              ]}
+            >
+              ?
+            </Popover>
+          </>
+        }
         description="Authors can post proposals regardless of their voting power."
         addrList={listAddrAuthors}
         onAddressChange={onAuthorAddressChange}
@@ -165,7 +141,7 @@ export default function StepTwo({
             }`}
             onClick={isStepValid ? () => moveToNextStep() : () => {}}
           >
-            Next: PROPOSALS & VOTING
+            Next: PROPOSAL & VOTING
           </button>
         </div>
       </div>
