+++ frontend/packages/client/src/components/index.js
@@ -35,3 +35,4 @@ export {
 export { default as FadeIn } from './FadeIn';
 export { default as UploadImageModal } from './UploadImageModal';
 export { default as Tablink } from './Tablink';
+export { default as Popover } from './Popover';
